requirements.txt
imageio==2.4.1
numpy==1.15.4
pycairo==1.16.3
Pygments==2.3.1
